degrade the maven-surefire-plugin to run all test without missing
