+++ sofa-serverless-runtime/arklet-springboot-starter/src/test/java/com/alipay/sofa/serverless/arklet/springboot/starter/SpringbootRunnerTest.java
@@ -14,27 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.serverless.arklet.spring;
+package com.alipay.sofa.serverless.arklet.spring.starter;
 
-import com.alipay.sofa.ark.common.util.ClassLoaderUtils;
-import com.alipay.sofa.ark.spi.constant.Constants;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
+import com.alipay.sofa.serverless.arklet.core.ArkletComponentRegistry;
+import org.junit.*;
 
 /**
- * @author Lunarscave
+ * @author mingmen
+ * @date 2023/6/26
  */
-public class SpringbootBaseTest {
+public class SpringbootRunnerTest extends SpringbootBaseTest {
 
-    @BeforeClass
-    public static void initSpringbootTest() {
-        ClassLoaderUtils.pushContextClassLoader(ClassLoader.getSystemClassLoader());
-        System.setProperty(Constants.EMBED_ENABLE, "true");
-        BaseSpringApplication.main(new String[] {});
-    }
-
-    @AfterClass
-    public static void destroySpringbootTest() {
-        System.setProperty(Constants.EMBED_ENABLE, "");
+    @Test
+    public void testArkletService() {
+        Assert.assertNotNull(ArkletComponentRegistry.getCommandServiceInstance());
+        Assert.assertNotNull(ArkletComponentRegistry.getHealthServiceInstance());
+        Assert.assertNotNull(ArkletComponentRegistry.getOperationServiceInstance());
     }
 }
