+++ sofa-serverless-runtime/arklet-springboot-starter/src/test/java/com/alipay/sofa/serverless/arklet/springboot/starter/health/endpoint/ArkHealthzEndpointTest.java
@@ -14,11 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.serverless.arklet.spring.health.endpoint;
+package com.alipay.sofa.serverless.arklet.spring.starter.health.endpoint;
 
 import com.alipay.sofa.serverless.arklet.core.health.model.Constants;
-import com.alipay.sofa.serverless.arklet.spring.SpringbootBaseTest;
-import com.alipay.sofa.serverless.arklet.spring.common.SpringbootUtil;
+import com.alipay.sofa.serverless.arklet.spring.starter.SpringbootBaseTest;
+import com.alipay.sofa.serverless.arklet.spring.starter.common.SpringbootUtil;
 import com.alipay.sofa.serverless.arklet.springboot.starter.health.endpoint.ArkHealthzEndpoint;
 import com.alipay.sofa.serverless.arklet.springboot.starter.health.endpoint.model.EndpointResponse;
 import com.alipay.sofa.serverless.arklet.springboot.starter.health.endpoint.model.EndpointResponseCode;
